So long, checkpoint
