+++ crates/sui-core/src/authority_active.rs
@@ -56,18 +56,11 @@ use tokio::time::Instant;
 pub mod gossip;
 use gossip::{gossip_process, GossipMetrics};
 
-pub mod checkpoint_driver;
-use crate::authority_active::checkpoint_driver::CheckpointMetrics;
 use crate::authority_client::NetworkAuthorityClientMetrics;
-use crate::epoch::reconfiguration::Reconfigurable;
-use checkpoint_driver::{checkpoint_process, get_latest_checkpoint_from_all, sync_to_checkpoint};
 
 pub mod execution_driver;
 
-use self::{
-    checkpoint_driver::CheckpointProcessControl,
-    execution_driver::{execution_process, ExecutionDriverMetrics},
-};
+use self::execution_driver::{execution_process, ExecutionDriverMetrics};
 
 // TODO: Make these into a proper config
 const MAX_RETRIES_RECORDED: u32 = 10;
@@ -269,38 +262,6 @@ where
             .clone()
     }
 
-    pub async fn sync_to_latest_checkpoint(self: Arc<Self>) -> SuiResult {
-        self.sync_to_latest_checkpoint_with_config(Default::default())
-            .await
-    }
-
-    pub async fn sync_to_latest_checkpoint_with_config(
-        self: Arc<Self>,
-        checkpoint_process_control: CheckpointProcessControl,
-    ) -> SuiResult {
-        let checkpoint_store = self.state.checkpoints.clone();
-
-        // TODO: fullnode should not get proposals
-        // TODO: potentially move get_latest_proposal_and_checkpoint_from_all and
-        // sync_to_checkpoint out of checkpoint_driver
-        let checkpoint_summary = get_latest_checkpoint_from_all(
-            self.agg_aggregator(),
-            checkpoint_process_control.extra_time_after_quorum,
-            checkpoint_process_control.timeout_until_quorum,
-        )
-        .await?;
-
-        let checkpoint_summary = match checkpoint_summary {
-            Some(c) => c,
-            None => {
-                info!(name = ?self.state.name, "no checkpoints found");
-                return Ok(());
-            }
-        };
-
-        sync_to_checkpoint(self, checkpoint_store, checkpoint_summary).await
-    }
-
     /// Spawn gossip process
     pub async fn spawn_gossip_process(self: Arc<Self>, degree: usize) -> JoinHandle<()> {
         // Number of tasks at most "degree" and no more than committee - 1
@@ -387,31 +348,5 @@ where
     /// Spawn pending certificate execution process
     pub async fn spawn_execute_process(self: Arc<Self>) -> JoinHandle<()> {
         spawn_monitored_task!(execution_process(self))
-    }
-}
-
-impl<A> ActiveAuthority<A>
-where
-    A: AuthorityAPI + Send + Sync + 'static + Clone + Reconfigurable,
-{
-    pub async fn spawn_checkpoint_process(
-        self: Arc<Self>,
-        metrics: CheckpointMetrics,
-    ) -> JoinHandle<()> {
-        self.spawn_checkpoint_process_with_config(CheckpointProcessControl::default(), metrics)
-            .await
-    }
-
-    pub async fn spawn_checkpoint_process_with_config(
-        self: Arc<Self>,
-        checkpoint_process_control: CheckpointProcessControl,
-        metrics: CheckpointMetrics,
-    ) -> JoinHandle<()> {
-        // Spawn task to take care of checkpointing
-        spawn_monitored_task!(checkpoint_process(
-            self,
-            &checkpoint_process_control,
-            metrics
-        ))
     }
 }
