+++ crates/sui-core/src/authority_active/gossip/configurable_batch_action_client.rs
@@ -5,9 +5,7 @@
 use crate::authority::AuthorityState;
 use crate::authority_aggregator::authority_aggregator_tests::*;
 use crate::authority_aggregator::{AuthAggMetrics, AuthorityAggregator};
-use crate::authority_client::{
-    AuthorityAPI, BatchInfoResponseItemStream, CheckpointStreamResponseItemStream,
-};
+use crate::authority_client::{AuthorityAPI, BatchInfoResponseItemStream};
 use crate::epoch::committee_store::CommitteeStore;
 use crate::safe_client::SafeClient;
 use async_trait::async_trait;
@@ -23,9 +21,8 @@ use sui_types::crypto::{get_key_pair, AuthorityKeyPair};
 use sui_types::error::SuiError;
 use sui_types::messages::{
     AccountInfoRequest, AccountInfoResponse, BatchInfoRequest, BatchInfoResponseItem,
-    CertifiedTransaction, CheckpointStreamRequest, CommitteeInfoRequest, CommitteeInfoResponse,
-    ObjectInfoRequest, ObjectInfoResponse, Transaction, TransactionInfoRequest,
-    TransactionInfoResponse,
+    CertifiedTransaction, CommitteeInfoRequest, CommitteeInfoResponse, ObjectInfoRequest,
+    ObjectInfoResponse, Transaction, TransactionInfoRequest, TransactionInfoResponse,
 };
 use sui_types::messages_checkpoint::{CheckpointRequest, CheckpointResponse};
 use sui_types::object::Object;
@@ -69,7 +66,6 @@ impl ConfigurableBatchActionClient {
         let state = AuthorityState::new_for_testing(
             committee,
             &secret,
-            None,
             None,
             None,
             tx_reconfigure_consensus,
@@ -200,14 +196,6 @@ impl AuthorityAPI for ConfigurableBatchActionClient {
         });
 
         Ok(Box::pin(tokio_stream::iter(items)))
-    }
-
-    async fn handle_checkpoint_stream(
-        &self,
-        request: CheckpointStreamRequest,
-    ) -> Result<CheckpointStreamResponseItemStream, SuiError> {
-        let stream = self.state.handle_checkpoint_streaming(request).await?;
-        Ok(Box::pin(stream))
     }
 
     async fn handle_checkpoint(
