+++ crates/sui-core/src/authority_client.rs
@@ -72,11 +72,6 @@ pub trait AuthorityAPI {
         request: CheckpointRequest,
     ) -> Result<CheckpointResponse, SuiError>;
 
-    async fn handle_checkpoint_stream(
-        &self,
-        request: CheckpointStreamRequest,
-    ) -> Result<CheckpointStreamResponseItemStream, SuiError>;
-
     async fn handle_committee_info_request(
         &self,
         request: CommitteeInfoRequest,
@@ -84,8 +79,6 @@ pub trait AuthorityAPI {
 }
 
 pub type BatchInfoResponseItemStream = BoxStream<'static, Result<BatchInfoResponseItem, SuiError>>;
-pub type CheckpointStreamResponseItemStream =
-    BoxStream<'static, Result<CheckpointStreamResponseItem, SuiError>>;
 
 #[derive(Clone)]
 pub struct NetworkAuthorityClient {
@@ -248,21 +241,6 @@ impl AuthorityAPI for NetworkAuthorityClient {
             .await
             .map(tonic::Response::into_inner)
             .map_err(Into::into)
-    }
-
-    /// Stream checkpoint notifications
-    async fn handle_checkpoint_stream(
-        &self,
-        request: CheckpointStreamRequest,
-    ) -> Result<CheckpointStreamResponseItemStream, SuiError> {
-        let stream = self
-            .client()
-            .checkpoint_info(request)
-            .await
-            .map(tonic::Response::into_inner)?
-            .map_err(Into::into);
-
-        Ok(Box::pin(stream))
     }
 
     async fn handle_committee_info_request(
@@ -469,14 +447,6 @@ impl AuthorityAPI for LocalAuthorityClient {
         state.handle_checkpoint_request(&request)
     }
 
-    async fn handle_checkpoint_stream(
-        &self,
-        request: CheckpointStreamRequest,
-    ) -> Result<CheckpointStreamResponseItemStream, SuiError> {
-        let stream = self.state.handle_checkpoint_streaming(request).await?;
-        Ok(Box::pin(stream))
-    }
-
     async fn handle_committee_info_request(
         &self,
         request: CommitteeInfoRequest,
@@ -496,7 +466,6 @@ impl LocalAuthorityClient {
             &secret,
             None,
             Some(genesis),
-            None,
             tx_reconfigure_consensus,
         )
         .await;
