+++ crates/sui-core/src/epoch/mod.rs
@@ -4,6 +4,8 @@
 pub mod committee_store;
 pub mod reconfiguration;
 
+/*
 #[cfg(test)]
 #[path = "./tests/reconfiguration_tests.rs"]
 mod reconfiguration_tests;
+*/
