+++ crates/sui-core/src/lib.rs
@@ -7,7 +7,6 @@ pub mod authority_aggregator;
 pub mod authority_batch;
 pub mod authority_client;
 pub mod authority_server;
-pub mod checkpoints;
 pub mod consensus_adapter;
 pub mod epoch;
 pub mod event_handler;
