+++ crates/sui-network/build.rs
@@ -88,16 +88,6 @@ fn main() -> Result<()> {
         )
         .method(
             Method::builder()
-                .name("checkpoint_info")
-                .route_name("FollowCheckpointStream")
-                .input_type("sui_types::messages::CheckpointStreamRequest")
-                .output_type("sui_types::messages::CheckpointStreamResponseItem")
-                .server_streaming()
-                .codec_path(codec_path)
-                .build(),
-        )
-        .method(
-            Method::builder()
                 .name("committee_info")
                 .route_name("CommitteeInfo")
                 .input_type("sui_types::messages::CommitteeInfoRequest")
