+++ crates/sui-types/src/crypto.rs
@@ -1354,11 +1354,6 @@ mod bcs_signable {
     impl BcsSignable for crate::committee::CommitteeWithNetAddresses {}
     impl BcsSignable for crate::messages_checkpoint::CheckpointSummary {}
     impl BcsSignable for crate::messages_checkpoint::CheckpointContents {}
-    impl BcsSignable for crate::messages_checkpoint::CheckpointProposalContents {}
-    impl BcsSignable for crate::messages_checkpoint::CheckpointProposalSummary {}
-    impl BcsSignable for crate::messages_checkpoint::CheckpointFragmentMessageHeader {}
-    impl BcsSignable for crate::messages_checkpoint::CheckpointFragmentMessageChunk {}
-    impl BcsSignable for crate::messages_checkpoint::CheckpointFragmentMessage {}
 
     impl BcsSignable for crate::messages::CommitteeInfoResponse {}
     impl BcsSignable for crate::messages::TransactionEffects {}
